refactor: extract client components from homepage
